+++ tests/local_onionshare_settings_dialog_no_tor_test.py
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+import json
+import unittest
+import time
+from PyQt5 import QtCore, QtTest
+
+from onionshare import strings
+from .SettingsGuiBaseTest import SettingsGuiBaseTest, OnionStub
+
+
+class SettingsGuiTest(unittest.TestCase, SettingsGuiBaseTest):
+    @classmethod
+    def setUpClass(cls):
+        cls.gui = SettingsGuiBaseTest.set_up()
+
+    @classmethod
+    def tearDownClass(cls):
+        SettingsGuiBaseTest.tear_down()
+
+    def test_gui_no_tor(self):
+        self.gui.onion = OnionStub(False)
+        self.gui.reload_settings()
+        self.run_settings_gui_tests()
+
+
+if __name__ == "__main__":
+    unittest.main()
